refactor(rust_core): split lib.rs into focused submodules
Preemptively splits the 486 LOC rust_core/src/lib.rs into three focused
submodules to stay below the monolith threshold:

- dynamics.rs: inverse_dynamics_batch + pure-Rust tests
- kinematics.rs: com_batch + pure-Rust tests
- interpolation.rs: interpolate_phases_rs + pure-Rust tests

lib.rs remains the public entrypoint, declaring the submodules,
re-exporting the three pyfunctions, and assembling the PyO3 module.
No function signatures, return types, or runtime behaviour changed; the
public PyO3 API (opensim_models_core.inverse_dynamics_batch / com_batch /
interpolate_phases_rs) is preserved exactly.

Fixes #127
